Changed the test files names.
