+++ .gitignore
@@ -280,7 +280,6 @@ prueba.cpp
 myjob.sh
 .myjob_*
 hola.sh
-test*
 generar_circuito.py
 
 modules_alvaro.sh
