+++ examples/CESGA_use_cases/Distributed_Grover/test_circuito_target.json
@@ -1,0 +1,44 @@
+{
+    "id": "241773_356102",
+    "config": {
+        "shots": 1000,
+        "method": "statevector",
+        "num_clbits": 1,
+        "num_qubits": 1,
+        "seed": 123123
+    },
+    "instructions": [
+        {
+            "name": "rcontrol",
+            "qubits": [],
+            "instructions": [
+                {
+                    "name": "cx",
+                    "qubits": [-1,0]
+                }
+            ],
+            "qpus": ["241773_356101"]
+        },
+        {
+            "name": "recv",
+            "qubits": [],
+            "remote_conditional_reg": [0],
+            "qpus": ["241773_356101"]
+        },
+        {
+            "name": "x",
+            "qubits": [0],
+            "remote_conditional_reg": [0],
+            "params": []
+        },
+        {
+            "name": "measure",
+            "qubits": [0],
+            "clbits": [0],
+            "clreg": []
+        }
+    ],
+    "sending_to": [],
+    "is_dynamic": true,
+    "has_cc": true
+}
